Oban improvements
- Only install in prod and dev
- Upgrade to 2.10 (RC for now)
- Use new resolver 'forbidden' feature
- Remove no-longer-needed plugins
- Only load ObanWeb code when installed
