+++ config/config.exs
@@ -41,11 +41,7 @@ config :changelog,
 config :changelog, Oban,
   repo: Changelog.Repo,
   queues: [audio_updater: 10, scheduled: 5, email: 1, feeds: 1],
-  plugins: [
-    Oban.Plugins.Pruner,
-    Oban.Plugins.Gossip,
-    Oban.Web.Plugins.Stats
-  ]
+  plugins: [Oban.Plugins.Pruner]
 
 config :changelog, Changelog.Mailer, adapter: Bamboo.LocalAdapter
