+++ mix.exs
@@ -1,6 +1,9 @@
 defmodule Changelog.Mixfile do
   use Mix.Project
 
+  # Change this to [:prod] if you don't have a web license available
+  @oban_envs [:dev, :prod]
+
   Code.compile_file("config/secret_or_env.exs")
 
   def project do
@@ -44,7 +47,7 @@ defmodule Changelog.Mixfile do
       {:phoenix_live_reload, "~> 1.2", only: :dev},
       {:plug_cowboy, "~> 2.5"},
       {:oban, "~> 2.15"},
-      {:oban_web, "~> 2.9", repo: "oban"},
+      {:oban_web, "~> 2.10.0-rc.2", repo: "oban", only: @oban_envs},
       {:timex, "~> 3.0"},
       {:scrivener_ecto, "~> 2.0"},
       {:scrivener_html, "~> 1.8", github: "jerodsanto/scrivener_html", branch: "phx-1-7"},
